Main Menu
added main menu that includes an options screen and a level select that can navigate to level 1 and sample scene (level 2 for now). If its not working for you, make sure that in file->BuildSettings that under: Scenes in Build, it you can see the MainMenu, level1, and SampleScene are all there and have a check mark. If they're not there, you have to drag and drop them to that box and put a check mark on them.