Save manager, weapon select functionalities
If the player has no save, they have to choose their weapon, otherwise they select their last saved level. Also note that in the level select menu, the on click events for the level select buttons has runtime as off because the save progress is still under development, there needs to be criteria in which designates whether the player has completed the level or not.